0017_zoff_tc___fix_tox_group_peer_count_and_tox_group_get_peerlist_functions

+++ toxcore/tox.h
@@ -4068,7 +4068,7 @@ typedef enum Tox_Err_Group_Peer_Query {
 } Tox_Err_Group_Peer_Query;
 
 
-size_t tox_group_peer_count(const Tox *tox, uint32_t group_number, Tox_Err_Group_Peer_Query *error);
+uint32_t tox_group_peer_count(const Tox *tox, uint32_t group_number, Tox_Err_Group_Peer_Query *error);
 
 void tox_group_get_peerlist(const Tox *tox, uint32_t group_number, uint32_t *peerlist, Tox_Err_Group_Peer_Query *error);
